@@ -1,18 +1,29 @@
 {
-  "name": "next-netlify-starter",
-  "version": "0.5.0",
-  "private": true,
+  "name": "nextjs-toolbox",
+  "version": "1.0.0",
+  "author": "Netlify Inc",
+  "contributors": [
+    "Tara Manicsic",
+    "Prince Wilson",
+    "Charlie Gerard",
+    "Ben Hong"
+  ],
+  "homepage": "https://github.com/netlify-templates/nextjs-toolbox",
+  "bugs": {
+    "url": "https://github.com/netlify-templates/nextjs-toolbox/issues"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "export": "next export"
   },
   "dependencies": {
-    "next": "^12.0.7",
+    "cypress": "^10.0.3",
+    "next": "^12.1.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@netlify/plugin-nextjs": "^4.0.0"
+    "@netlify/plugin-nextjs": "^4.3.2"
   }
 }
